client/requirements.txt
# client/requirements.txt

paramiko==3.5.0
web3==6.4.0
python-dotenv==1.0.0
eth-account==0.8.0
cryptography==40.0.1
